separate on modules slider and scale
