+++ js/constants.js
@@ -1,0 +1,70 @@
+const SCALE_STEP = 25;
+const SCALE_MIN = 25;
+const SCALE_MAX = 100;
+const SCALE_DEFAULT = 100;
+const EFFECTS = [
+  {
+    name: 'none',
+    style: 'none',
+    min: 0,
+    max: 100,
+    step: 1,
+    unit: '',
+  },
+  {
+    name: 'chrome',
+    style: 'grayscale',
+    min: 0,
+    max: 1,
+    step: 0.1,
+    unit: '',
+  },
+  {
+    name: 'sepia',
+    style: 'sepia',
+    min: 0,
+    max: 1,
+    step: 0.1,
+    unit: ''
+  },
+  {
+    name: 'marvin',
+    style: 'invert',
+    min: 0,
+    max: 100,
+    step: 1,
+    unit: '%',
+  },
+  {
+    name: 'phobos',
+    style: 'blur',
+    min: 0,
+    max: 3,
+    step: 0.1,
+    unit: 'px',
+  },
+  {
+    name: 'heat',
+    style: 'brighthess',
+    min: 1,
+    max: 3,
+    step: 0.1,
+    unit: '',
+  },
+];
+const EFFECT_DEFAULT = EFFECTS[0];
+const MAX_HASHTAG_COUNT = 5;
+const VALID_SYMBOLS = /^#[a-zа-я0-9]{1,19}$/i;
+const TAG_ERROR_TEXT = 'Хештеги введёны неверно.';
+
+export {
+  SCALE_STEP,
+  SCALE_MIN,
+  SCALE_MAX,
+  SCALE_DEFAULT,
+  EFFECTS,
+  EFFECT_DEFAULT,
+  MAX_HASHTAG_COUNT,
+  VALID_SYMBOLS,
+  TAG_ERROR_TEXT
+};
