+++ js/main.js
@@ -1,6 +1,6 @@
 import './form.js';
-import './scale.js';
-import './effect.js';
+import './scale/scale.js';
+import './slider/effect.js';
 import { renderGallery } from './gallery.js';
 import { AVATAR_NUMBER_MAX, AVATAR_NUMBER_MIN, COMMENTS, LIKES_AMOUNT_MAX, LIKES_AMOUNT_MIN, NAMES, getPhotos, PHOTOS_AMOUNT, PHOTO_DESCRIPTIONS } from './data.js';
