+++ js/scale/reset-scale.js
@@ -1,0 +1,6 @@
+import { scaleImage } from './scale-image.js';
+import { SCALE_DEFAULT } from '../constants.js';
+
+const resetScale = () => scaleImage(SCALE_DEFAULT);
+
+export { resetScale };
