+++ js/slider/close-slider.js
@@ -1,0 +1,7 @@
+const sliderLevelContainer = document.querySelector('.img-upload__overlay');
+
+const hideSlider = () => {
+  sliderLevelContainer.classList.add('hidden');
+};
+
+export { hideSlider };
