+++ js/slider/open-slider.js
@@ -1,0 +1,7 @@
+const sliderLevelContainer = document.querySelector('.img-upload__overlay');
+
+const showSlider = () => {
+  sliderLevelContainer.classList.remove('hidden');
+};
+
+export { showSlider };
